A few fixes and reduce the header includes.

+++ Projects/CoX/Common/Servers/InternalEvents.h
@@ -190,7 +190,7 @@ struct ClientConnectedData
     uint64_t m_session;
     uint32_t m_server_id;     // id of the server the client connected to.
     uint32_t m_sub_server_id; // only used when server_id is the map server
-    uint32_t m_char_db_id;       // id of the character connecting
+    uint64_t m_char_db_id;       // id of the character connecting
     template<class Archive>
     void serialize(Archive &ar)
     {
@@ -203,7 +203,7 @@ ONE_WAY_MESSAGE(Internal_EventTypes,ClientConnected)
 struct ClientDisconnectedData
 {
     uint64_t m_session;
-    uint32_t m_char_db_id;       // id of the character disconnected
+    uint64_t m_char_db_id;       // id of the character disconnected
     template<class Archive>
     void serialize(Archive &ar)
     {
